fix: explicitly check for null before attempting to deactivate custom You are Fired subtext
We cannot use null propagation because, in the event the game object was destroyed by Unity, Unity does not actually set the object to null, but instead sets an internal flag which null propagation cannot detect.

Fixes a NullPointerException thrown when leaving leaving a lobby in the middle of the fired cutscene.